# internal/api/openapi.yaml
openapi: "3.1.0"
info:
  title: Vedetta NVR API
  version: "0.1.0"
  description: API for managing cameras, events, recordings, and object detection in the Vedetta Network Video Recorder.
servers:
  - url: /
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /api/cameras/{name}/doorbell/{event_id}/answer:
    post:
      operationId: answerDoorbell
      summary: Acknowledge (answer) a doorbell ring
      tags: [cameras]
      parameters:
        - name: name
          in: path
          required: true
          schema: { type: string }
        - name: event_id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Ring acknowledged (idempotent; returns the first answerer)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoorbellAnswerResponse"
        "404":
          description: Ring not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    DoorbellAnswerResponse:
      type: object
      required: [event_id, answered_at, answered_by]
      properties:
        event_id: { type: string }
        answered_at: { type: string, format: date-time }
        answered_by: { type: string }
...